+++ src/lib/components/basic/icons/activeCopy.svelte
