+++ src/lib/components/basic/icons/add.svelte
