+++ src/lib/components/basic/icons/binIcon.svelte
