+++ src/lib/components/basic/icons/blueClose.svelte
