+++ src/lib/components/basic/icons/closePanel.svelte
