+++ src/lib/components/basic/icons/closedEye.svelte
