+++ src/lib/components/basic/icons/copyIcon.svelte
