+++ src/lib/components/basic/icons/editIcon.svelte
