+++ src/lib/components/basic/icons/eye.svelte
