+++ src/lib/components/basic/icons/groupIcon.svelte
