+++ src/lib/components/basic/icons/infoIcon.svelte
