+++ src/lib/components/basic/icons/key.svelte
