+++ src/lib/components/basic/icons/lens.svelte
