+++ src/lib/components/basic/icons/locked.svelte
