+++ src/lib/components/basic/icons/logo.svelte
