+++ src/lib/components/basic/icons/maximize.svelte
