+++ src/lib/components/basic/icons/more.svelte
