+++ src/lib/components/basic/icons/rightArrow.svelte
