+++ src/lib/components/basic/icons/sensitiveEyeBlue.svelte
