+++ src/lib/components/basic/icons/share.svelte
