+++ src/lib/components/basic/icons/tick.svelte
