+++ src/lib/components/dashboard/groups/GroupSideList.svelte
@@ -9,7 +9,7 @@
     adminStatus,
   } from "../store";
   import { Group } from "../dtos";
-  import Add from "../../basic/add.svelte";
+  import Add from "../../basic/icons/add.svelte";
 
   let iconColor = "#6E7681"; //sheffieldgrey:
