+++ src/lib/components/dashboard/icons.ts
@@ -1,25 +1,25 @@
 
-export { default as Locked } from "../basic/locked.svelte";
-export { default as Eye } from "../basic/eye.svelte";
-export { default as Unlocked } from "../basic/unlocked.svelte";
-export { default as CopyIcon } from "../basic/copyIcon.svelte";
-export { default as ActiveCopy } from "../basic/activeCopy.svelte";
-export { default as ClosedEye } from "../basic/closedEye.svelte";
-export { default as BinIcon } from "../basic/binIcon.svelte";
-export { default as Lens } from "../basic/lens.svelte";
-export { default as ClosePanel } from "../basic/closePanel.svelte";
-export { default as InfoIcon } from "../basic/infoIcon.svelte";
-export { default as DownArrow } from "../basic/downArrow.svelte";
-export { default as RightArrow } from "../basic/rightArrow.svelte";
-export { default as More } from "../basic/more.svelte";
-export { default as SensitiveEye } from "../basic/sensitiveEye.svelte";
-export { default as SensitiveEyeBlue } from "../basic/sensitiveEyeBlue.svelte";
-export { default as Share } from "../basic/share.svelte";
-export { default as Add } from "../basic/add.svelte";
-export { default as Profile } from "../basic/profile.svelte";
-export { default as Logo } from "../basic/logo.svelte";
+export { default as Locked } from "../basic/icons/locked.svelte";
+export { default as Eye } from "../basic/icons/eye.svelte";
+export { default as Unlocked } from "../basic/icons/unlocked.svelte";
+export { default as CopyIcon } from "../basic/icons/copyIcon.svelte";
+export { default as ActiveCopy } from "../basic/icons/activeCopy.svelte";
+export { default as ClosedEye } from "../basic/icons/closedEye.svelte";
+export { default as BinIcon } from "../basic/icons/binIcon.svelte";
+export { default as Lens } from "../basic/icons/lens.svelte";
+export { default as ClosePanel } from "../basic/icons/closePanel.svelte";
+export { default as InfoIcon } from "../basic/icons/infoIcon.svelte";
+export { default as DownArrow } from "../basic/icons/downArrow.svelte";
+export { default as RightArrow } from "../basic/icons/rightArrow.svelte";
+export { default as More } from "../basic/icons/more.svelte";
+export { default as SensitiveEye } from "../basic/icons/sensitiveEye.svelte";
+export { default as SensitiveEyeBlue } from "../basic/icons/sensitiveEyeBlue.svelte";
+export { default as Share } from "../basic/icons/share.svelte";
+export { default as Add } from "../basic/icons/add.svelte";
+export { default as Profile } from "../basic/icons/profile.svelte";
+export { default as Logo } from "../basic/icons/logo.svelte";
 export { default as Toggle } from "../basic/toggle.svelte";
-export { default as GroupIcon } from "../basic/groupIcon.svelte";
-export { default as UserIcon } from "../basic/userIcon.svelte";
-export { default as BlueClose } from "../basic/blueClose.svelte";
-export { default as EditIcon } from "../basic/editIcon.svelte";
+export { default as GroupIcon } from "../basic/icons/groupIcon.svelte";
+export { default as UserIcon } from "../basic/icons/userIcon.svelte";
+export { default as BlueClose } from "../basic/icons/blueClose.svelte";
+export { default as EditIcon } from "../basic/icons/editIcon.svelte";
