+++ src/lib/components/popup/icons.ts
@@ -1,13 +1,13 @@
 
-export { default as Maximize } from "../basic/maximize.svelte";
-export { default as Lens } from "../basic/lens.svelte";
-export { default as ClosePanel } from "../basic/closePanel.svelte";
-export { default as Add } from "../basic/add.svelte";
-export { default as DownArrow } from "../basic/downArrow.svelte";
-export { default as RightArrow } from "../basic/rightArrow.svelte";
-export { default as ActiveCopy } from "../basic/activeCopy.svelte";
-export { default as Locked } from "../basic/locked.svelte";
-export { default as Eye } from "../basic/eye.svelte";
-export { default as Unlocked } from "../basic/unlocked.svelte";
-export { default as CopyIcon } from "../basic/copyIcon.svelte";
-export { default as ClosedEye } from "../basic/closedEye.svelte";
+export { default as Maximize } from "../basic/icons/maximize.svelte";
+export { default as Lens } from "../basic/icons/lens.svelte";
+export { default as ClosePanel } from "../basic/icons/closePanel.svelte";
+export { default as Add } from "../basic/icons/add.svelte";
+export { default as DownArrow } from "../basic/icons/downArrow.svelte";
+export { default as RightArrow } from "../basic/icons/rightArrow.svelte";
+export { default as ActiveCopy } from "../basic/icons/activeCopy.svelte";
+export { default as Locked } from "../basic/icons/locked.svelte";
+export { default as Eye } from "../basic/icons/eye.svelte";
+export { default as Unlocked } from "../basic/icons/unlocked.svelte";
+export { default as CopyIcon } from "../basic/icons/copyIcon.svelte";
+export { default as ClosedEye } from "../basic/icons/closedEye.svelte";
